\name{AASP}
\alias{AASP}
\docType{data}
\title{
Academic Achievment Score Projection
}
\description{
Data from a thesis about academic achievement Projection -aasp-, with 632 rows and 205 columns.

}
\usage{data("AASP")}


\source{
https://github.com/cgamboasanabria/dCUR/tree/master/inst/AASP_description_data.xlsx
}

\examples{
data(AASP)
summary(AASP)
}
