\name{AASP}
\alias{AASP}
\docType{data}
\title{
Academic Achievement Score Projection -AASP-
}
\description{
Data from a Ph.D. thesis about Academic Achievement Score Projection, with 632 rows and 205 columns.

}
\usage{data("AASP")}


\source{
Further information about variables can be found in this link:

https://github.com/cgamboasanabria/dCUR/tree/master/inst/AASP_description_data.xlsx
}

\examples{
data(AASP)
dim(AASP)
}
