% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevant_variables_plot.R
\name{relevant_variables_plot}
\alias{relevant_variables_plot}
\title{relevant_variables_plot}
\usage{
relevant_variables_plot(data)
}
\arguments{
\item{data}{An object resulting from a call to CUR.}
}
\description{
\code{relevant_variables_plot} returns a bar graph which contains the leverages of the most relevant variable of data matrix according to CUR decomposition.
}
\details{
Relevant Variables Plot
}
\examples{
\donttest{
result <- CUR(data=AASP, variables=hoessem:notabachillerato,
k=20, rows = 1, columns = .2, standardize = TRUE,
cur_method = "sample_cur")
relevant_variables_plot(result)
}

}
\references{
\insertRef{dynamyCUR}{dCUR}
}
\seealso{
\code{\link[dCUR]{dCUR}}
\code{\link[dCUR]{CUR}}
}
\author{
Cesar Gamboa-Sanabria, Stefany Matarrita-Munoz, Katherine Barquero-Mejias, Greibin Villegas-Barahona, Mercedes Sanchez-Barba and Maria Purificacion Galindo-Villardon.
}
